refactor: extract YARA and IOC workers into runtime module
Move worker loop implementations out of src/main.rs into
src/runtime/yara.rs and src/runtime/ioc.rs so main.rs is
reduced to orchestration and entrypoint selection.

- runtime/yara.rs: alert builders, file worker, memory worker
- runtime/ioc.rs: IOC hash worker
- main.rs now calls spawn_yara_file_worker, spawn_yara_memory_worker,
  and spawn_ioc_hash_worker instead of owning the loops directly

Behavior, alert schema, and configuration defaults unchanged.
