+++ src/runtime/mod.rs
@@ -1,1 +1,3 @@
+pub mod ioc;
 pub mod logging;
+pub mod yara;
